+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Change Log
+
+## 0.5.0
+
+Added `Maybe`, `OR`, `ErrorEntry` class implementations.  
+Added Extensions for `Collections`, `Enumerables`, `Dictionaries`, `Maybe`, `Types`.
+Added Responses.